+++ www/go/modules/community/addressbook/views/extjs3/BirthdaysPortlet.js
@@ -57,15 +57,13 @@ go.modules.community.addressbook.BirthdaysPortlet = Ext.extend(go.grid.GridPanel
 				sortable: false,
 				dataIndex: "birthday",
 				renderer: function(v, meta, record) {
-
-					var bday = "";
+					let bdate = "";
 					record.data.dates.forEach(function(date) {
-						if(date.type == "birthday") {
-							bday = date.date;
+						if(date.type === "birthday") {
+							bdate = date.date;
 						}
 					});
-
-					return go.util.Format.date(bday);
+					return go.util.Format.date(bdate);
 				}
 			}, {
 				header: t("Age"),
@@ -73,7 +71,17 @@ go.modules.community.addressbook.BirthdaysPortlet = Ext.extend(go.grid.GridPanel
 				sortable: false,
 				width: 100,
 				renderer: function(v, meta, record) {
-					return v + 1;
+					let bdate = "";
+					record.data.dates.forEach(function(date) {
+						if(date.type === "birthday") {
+							bdate = date.date;
+						}
+					});
+					const today = new Date(), bdt = new Date(bdate);
+					if(today.getUTCDate() !== bdt.getUTCDate() || today.getUTCMonth() !== bdt.getUTCMonth()) {
+						v++
+					}
+					return v;
 				}
 			}];
 
@@ -102,6 +110,11 @@ go.modules.community.addressbook.BirthdaysPortlet = Ext.extend(go.grid.GridPanel
 		go.modules.community.addressbook.BirthdaysPortlet.superclass.afterRender.call(this);
 
 		this.store.load();
+	},
+
+	getBirthDate: function(record) {
+
+		return bdate;
 	}
 });